+++ changelog.md
@@ -2,6 +2,7 @@
 --
 * When enabled, you can add images to articles from the assets module.
 * Bugfix: the sortable scope for the ```Image``` wasn't properly set.
+* Images are now enabled by default for articles.
 
 
 5.3.1 - 2017-03-07